pyproject.toml
[project]
name = "deploy-lab"
version = "0.1.0"
description = "Personal reference for application deployments on Linux. "
authors = [{ name = "Sérgio Dantas", email = "sergiodnts828@gmail.com" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.3.3",
]

[dependency-groups]
dev = ["ruff>=0.15.2"]

[tool.ruff.lint]
ignore = ['F401']

[tool.ruff.format]
quote-style = "single"
line-ending = "lf"
